sched: add period support for -deadline tasks.
Make it possible to specify a period (different or equal than deadline) for
-deadline tasks. Relative deadlines (D_i) are used on task arrivals to generate
new scheduling (absolute) deadlines as "d = t + D_i", and periods (P_i) to
postpone the scheduling deadlines as "d = d + P_i" when the budget is zero.
This is in general useful to model (and schedule) tasks that have slow
activation rates (long periods), but have to be scheduled soon once activated
(short deadlines).

Signed-off-by: Harald Gustafsson <harald.gustafsson@ericsson.com>
Signed-off-by: Dario Faggioli <raistlin@linux.it>
Signed-off-by: Juri Lelli <juri.lelli@gmail.com>