sched: -dl AC, compare span and actual task affinity
This resolves a bug that caused sched_setscheduler2 to fail. We
enforce -dl tasks to have affinity equal to the cpus of the
root_domain they run on. However, we compared the span with
cpus_allowed, that is statically initialized with CPU_MASK_ALL.
We now perform an and between cpus_allowed and cpu_active_mask
to compute the actual task affinity and be able to compare this
with the span.