sched: update rq clock inside dl_task_timer
While thoroughly testing SCHED_DEADLINE, we noticed a small glitch when
SCHED_DEADLINE was running on a NOHZ=y system. Analyzing the problem,
we found that at the time replenishment timer callback fires (dl_task_timer()
function is called), tsc and gtod values sometimes result to be highly
unconsistent. This means that we get an unconsistent value for exec_start
of the just replenished -deadline task, that it is throttled right back and
put to sleep without executing.

Differently from a "normal" enqueue, we enqueue back a throttled -deadline
task inside the timer callback, and it is not certain if a scheduler_tick()
occurred any close to this moment. Since sched_tick() calls sched_clock_tick()
to fix things between tsc and gtod (and to then update rq->clock), we
now do the same inside dl_task_timer() and this fixes the problem we were
experiencing.

Kudos to Fabio Checconi for helping debugging the thing :).

Signed-off-by: Juri Lelli <juri.lelli@gmail.com>